support parsing GQL ints as Reason ints

+++ src/root.ts
@@ -39,7 +39,7 @@ const writeObjectTypeDef = (type: Type) => {
 
 const scalarMap: ScalarMap = {
   String: "string",
-  Int: "float",
+  Int: "int",
   Float: "float",
   Boolean: "bool",
   ID: "string"
